+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## master
+
+* [feature] user is able to access MongoDB console
+* [feature] user is able to access redis-cli
+
 ## 0.3.2 / 2013-06-28
 
 * [improvement] API changes to handle user virtual server interactions